[refactor]: Split test by test type
